Remove test schedule source
