--- LightTimetable/Models/ScheduleSources/TestScheduleSettings.cs
@@ -1,7 +1,0 @@
-namespace LightTimetable.Models.ScheduleSources
-{
-    public sealed class TestScheduleSettings : VnzOsvitaSettings
-    {
-        public TestScheduleSettings() => UniversityId = 99999; 
-    }
-}
